updated header includes using the latest sfdc 1.8 version which comes with fixed MOS inline macro support. Also added own <ppcinline/macros.h> header file with support for our special function pointer type macros.

+++ openssl/crypto/bio/bio_lcl.h
@@ -72,7 +72,7 @@ union bio_addr_st {
 typedef unsigned int socklen_t;
 # endif
 # if defined(OPENSSL_SYS_AMIGA) && defined(__MORPHOS__)
-typedef unsigned long socklen_t;
+typedef int socklen_t;
 typedef unsigned long in_addr_t;
 # endif
